Core: extract package-modification primitives + TVApp channel injector (#445)
## What

Foundational refactor for issue #444 (mobile custom WGT + package
modification). Moves the desktop's portable package-editing logic into
`Apps2Samsung.Core` so the mobile head can reuse it. **No user-facing
change.**

Moved into Core (kept in their existing namespaces, so the desktop needs
no `using` changes):
- `PackageWorkspace` — extract/edit/repack a `.wgt`
- `IPackagePatcher` + `InstallResult`

Added to Core:
- `TvAppChannelInjector` — **settings-agnostic**: parses a `{name,url}`
JSON list and rewrites the `var channels = [...]` array in a TVApp wgt's
`js/main.js`. Channels are passed in, so the desktop feeds it
`AppSettings.TvAppChannelsJson` and mobile will feed its own settings.

Desktop `TvAppPackagePatcher` now delegates to the Core injector
(behavior unchanged).

## Verified
- Desktop (`Apps2Samsung.csproj`) builds clean.
- Mobile (`net10.0-android`) builds clean.

## Next (follow-up PRs)
- Extract the self-update *check* logic into Core (the download/apply
stays platform-specific).
- Wire the mobile features: custom-WGT file picker, TVApp channel
settings/UI, update check.
