+++ Apps2Samsung.Core/Models/InstallResult.cs
@@ -1,9 +1,9 @@
-﻿namespace Apps2Samsung.Models
+namespace Apps2Samsung.Models
 {
     public class InstallResult
     {
         public bool Success { get; init; }
-        public string ErrorMessage { get; init; }
+        public string ErrorMessage { get; init; } = string.Empty;
 
         public static InstallResult SuccessResult() => new() { Success = true };
         public static InstallResult FailureResult(string error) => new()
